exception files reorganisation
