+++ backend/src/main/java/marketplace/backend/exception/user/exceptions/MyEntityNotFoundException.java
@@ -1,4 +1,4 @@
-package marketplace.backend.exception.user;
+package marketplace.backend.exception.user.exceptions;
 
 public class MyEntityNotFoundException extends RuntimeException {
